backend/requirements.txt
django~=5.1.0
djangorestframework~=3.15.2
django-cors-headers~=4.4.0
channels~=4.1.0
daphne~=4.1.2
apscheduler~=3.10.4
